address comments from Tenzin and Greg
